+++ src/libs/Battle_interface/sea/battle_shipcommand.cpp
@@ -288,7 +288,7 @@ long BIShipCommandList::CommandAdding()
     auto *pAttr = m_pARoot->GetAttributeClass("Commands");
     if (!pAttr)
         return 0;
-    const long attrQuant = pAttr->GetAttributesNum();
+    const size_t attrQuant = pAttr->GetAttributesNum();
 
     for (long i = 0; i < attrQuant; i++)
     {
@@ -345,7 +345,7 @@ long BIShipCommandList::UserIconsAdding()
     auto *pAttr = m_pARoot->GetAttributeClass("UserIcons");
     if (!pAttr)
         return 0;
-    const long attrQuant = pAttr->GetAttributesNum();
+    const size_t attrQuant = pAttr->GetAttributesNum();
 
     for (long i = 0; i < attrQuant; i++)
     {
@@ -371,7 +371,7 @@ long BIShipCommandList::AbilityAdding()
     auto *pAttr = m_pARoot->GetAttributeClass("AbilityIcons");
     if (!pAttr)
         return 0;
-    const long attrQuant = pAttr->GetAttributesNum();
+    const size_t attrQuant = pAttr->GetAttributesNum();
 
     for (long i = 0; i < attrQuant; i++)
     {
